moving methods from InternalDataIncrementalIndex to OffheapOakIncrementalIndex
